pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.42",
    "werkzeug>=3.1.3",
    "flask-mail>=0.10.0",
    "stripe>=12.4.0",
    "razorpay>=1.4.2",
    "markdown>=3.8.2",
    "flask-socketio>=5.5.1",
    "eventlet>=0.40.2",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "requests>=2.32.4",
    "python-dotenv>=1.1.1",
    "bleach>=6.2.0",
    "pillow>=11.3.0",
    "django>=5.2.5",
    "django-crispy-forms>=2.4",
    "channels>=4.3.1",
    "celery>=5.5.3",
    "django-cors-headers>=4.7.0",
    "whitenoise>=6.9.0",
    "crispy-bootstrap5>=2025.6",
    "channels-redis>=4.3.0",
    "redis>=6.3.0",
]
